+++ log-processor/src/test/java/core/log/EnvConfTest.java
@@ -1,7 +1,6 @@
 package core.log;
 
 import core.framework.test.EnvResourceValidator;
-import core.framework.test.EnvWebValidator;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -13,10 +12,5 @@ class EnvConfTest {
     @Test
     void validateEnvResource() throws IOException {
         new EnvResourceValidator().validate();
-    }
-
-    @Test
-    void validateEnvWeb() throws IOException {
-        new EnvWebValidator().validate();
     }
 }