+++ core-ng/src/main/java/core/framework/http/ContentType.java
@@ -20,8 +20,8 @@ public final class ContentType {
     public static final ContentType APPLICATION_JSON = create("application/json", Charsets.UTF_8);
     public static final ContentType APPLICATION_JAVASCRIPT = create("application/javascript", Charsets.UTF_8);
     public static final ContentType APPLICATION_OCTET_STREAM = create("application/octet-stream", null);
-    public static final ContentType IMAGE_PNG = create("image/png", null);
     public static final ContentType APPLICATION_X_WWW_FORM_URLENCODED = create("application/x-www-form-urlencoded", Charsets.UTF_8);
+    public static final ContentType IMAGE_PNG = create("image/png", null);
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ContentType.class);